Modernize packaging to use `pyproject.toml` metadata + bump to Sopel 8+
The existing release of `sopel-meetbot` will continue to work just fine
for Sopel 7.x instances. Future development can focus on 8+ (and live in
Python-3-only land).
