+++ NEWS
@@ -1,4 +1,5 @@
-Version 1.0.0
-=============
+## Changelog
+
+### 1.0.0
 
 Initial release of `meetbot.py` from Sopel 7.1.x as its own package.
