--- setup.cfg
@@ -1,25 +1,0 @@
-[metadata]
-name = sopel-meetbot
-version = 1.0.0
-description = The meetbot plugin formerly shipped with Sopel prior to version 8.0.
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/sopel-irc/sopel-meetbot
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.0,<9
-
-[options.entry_points]
-sopel.plugins =
-    meetbot = sopel_meetbot
